+++ pages/the-cast.tsx
@@ -5,7 +5,8 @@ import Image from 'next/image'
 import { WaveBackground } from '../components/WaveBackground'
 import { Navbar } from '../components/Navbar'
 import { Footer } from '../components/Footer'
-import { PhotoCollage, AllInOne } from '../components/TheInterviewWeNeverSaw'
+import { OrnateFrame } from '../components/OrnateFrame'
+import { YearbookImage } from '../components/YearbookImage'
 
 const imgVariants = {
   initial: { opacity: 0, y: 10, skew: 1 },
@@ -34,7 +35,7 @@ export default function Yearbook({ characters, theme, videos }) {
   const y1 = useTransform(scrollY, [0, 500], [10, 50])
   const y2 = useTransform(scrollY, [0, 300], [0, -100])
   const y3 = useTransform(scrollY, [100, 600], [0, -200])
-
+  const y4 = useTransform(scrollY, [100, 600], [0, -200])
   return (
     <>
       <div className="bg-wall">
@@ -50,15 +51,9 @@ export default function Yearbook({ characters, theme, videos }) {
               />
             </div>
 
-            <motion.h1
-              className="font-PlayfairDisplay text-6xl pl-0 leading-normal font-bold text-popstar-hover text-center mx-2"
-              layoutId="h1"
-            >
-              The Interview We Never Saw
-            </motion.h1>
-            <h1 className="text-2xl mt-10 font-AveriaSerifLibre text-popstar-hover font-extrabold text-center">
-              starring
-            </h1>
+            <div className="flex flex-col justify-center ">
+              <OrnateFrame label="The Cast" color={theme.primary} />
+            </div>
 
             <div className="w-full ">
               <WaveBackground />
@@ -78,16 +73,53 @@ export default function Yearbook({ characters, theme, videos }) {
           </div>
         </div>
         <motion.div className="container mx-auto relative z-10">
-          <div className="relative z-50 bg-white  py-2 my-20 w-5/6 md:w-1/2 mx-auto shadow-md border-jasmine-faded border-8 ">
-            <p className="text-lg text-black text-center font-NotoSerif px-2">
-              The Meghan, Harry and Boris impressions came about via whimsical
-              enactment: I was with a friend, walking our dogs on a golf course
-              in Sunningdale, Berkshire, when I pretended to conduct a mock
-              interview: the premise was born.
+          <div className="min-h-screen flex flex-col items-center justify-center mb-10">
+            <motion.div
+              className="absolute right-0 top-14 invisible md:visible"
+              style={{ y: y1, x: 0 }}
+            >
+              <Image
+                src="/svg/frame-5.svg"
+                width="112px"
+                height="134px"
+                alt="Ornate Frame"
+              />
+            </motion.div>
+            <motion.div
+              className="absolute right-36 top-0 scale-90 invisible md:visible"
+              style={{ y: y2, x: 0 }}
+            >
+              <Image
+                src="/svg/frame-6.svg"
+                width="91px"
+                height="102px"
+                alt="Ornate Frame"
+              />
+            </motion.div>
+            <div className="flex flex-col mt-12">
+              {characters &&
+                characters.map((character, index) => {
+                  console.log('index: ', index)
+                  return (
+                    <>
+                      {index % 2 === 0 ? (
+                        <LeftsideProfile character={character} key={index} />
+                      ) : (
+                        <RightsideProfile character={character} key={index} />
+                      )}
+                    </>
+                  )
+                })}
+            </div>
+
+            <p className="border-2 p-8 font-PlayfairDisplay text-ocean text-lg text-bold  rounded-xl border-golden">
+              <span className="text-3xl text-golden">*</span> it&lsquo;s perhaps
+              worth pointing out in this day and age that these quotes are part
+              of my parody video scripts and in no way were any real life
+              individuals close to resembling these levels of honesty and
+              transparency in their respective dialogues.
             </p>
           </div>
-          {/* <PhotoCollage characters={characters} /> */}
-          <AllInOne characters={characters} />
         </motion.div>
         <Footer
           theme={theme.footer}
@@ -98,6 +130,93 @@ export default function Yearbook({ characters, theme, videos }) {
         />
       </div>
     </>
+  )
+}
+
+const LeftsideProfile = ({ character, key }) => {
+  return (
+    <div key={key} className="flex flex-col lg:flex-row items-center mb-32">
+      <motion.div
+        className="flex flex-col md:flex-row lg:w-1/3 items-center my-4"
+        initial={{ opacity: 0 }}
+        animate={{
+          opacity: 1,
+          transition: { duration: 1, delay: 1 }
+        }}
+        exit={{ opacity: 0, transition: { duration: 1, delay: 1 } }}
+      >
+        <div className="order-2 md:order-1 lg:order-2">
+          <YearbookImage character={character} key={character.id} />
+        </div>
+        <div className="flex order-1 flex-row ml-4 lg:mr-3 items-center p-4 bg-moss-green bg-opacity-50 rounded-t-3xl rounded-br-3xl md:w-1/3 w-1/2 lg:rounded-bl-3xl lg:rounded-br-none ">
+          <p className="font-AveriaSerifLibre">
+            <span className="text-4xl">&quot;</span>
+            <span className="text-xl ">{character.quote}</span>
+            <span className="text-4xl leading-5">&quot;</span>
+          </p>
+        </div>
+      </motion.div>
+      <div className="lg:pl-10 w-9/12 sm:w-full px-3 flex lg:w-2/3 flex-col items-start justify-center ">
+        <motion.h1
+          className="text-4xl text-left font-PlayfairDisplay "
+          layoutId="h1"
+          initial={{ scale: 0.9 }}
+          animate={{ scale: 1 }}
+          transition={{ duration: 0.2 }}
+        >
+          {character.name}
+        </motion.h1>
+        <motion.p className="text-left text-4xl my-4 font-AlexBrush">
+          {character.quip}
+        </motion.p>
+        <motion.p className="text-left text-lg font-NotoSerif">
+          {character.about}
+        </motion.p>
+      </div>
+    </div>
+  )
+}
+const RightsideProfile = ({ character, key }) => {
+  return (
+    <div key={key} className="flex flex-col lg:flex-row items-center mb-32">
+      <div className="lg:pl-10 w-9/12 sm:w-full px-3 flex lg:w-2/3 flex-col items-start justify-center order-2 lg:order-1">
+        <motion.h1
+          className="text-4xl text-left font-PlayfairDisplay"
+          layoutId="h1"
+          initial={{ scale: 0.9 }}
+          animate={{ scale: 1 }}
+          transition={{ duration: 0.2 }}
+        >
+          {character.name}
+        </motion.h1>
+        <motion.p className="text-left text-4xl my-4 font-AlexBrush">
+          {character.quip}
+        </motion.p>
+        <motion.p className="text-left text-lg font-NotoSerif">
+          {character.about}
+        </motion.p>
+      </div>
+      <motion.div
+        className="flex flex-col md:flex-row lg:w-1/3 items-center justify-end my-4 order-1 lg:order-2 "
+        initial={{ opacity: 0 }}
+        animate={{
+          opacity: 1,
+          transition: { duration: 1, delay: 1 }
+        }}
+        exit={{ opacity: 0, transition: { duration: 1, delay: 1 } }}
+      >
+        <div className="order-2 ">
+          <YearbookImage character={character} key={character.id} />
+        </div>
+        <div className="flex order-1 flex-row mr-4 lg:mr-3 items-center p-4 text-left bg-jasmine-faded rounded-t-3xl rounded-bl-3xl md:w-1/3 w-1/2 lg:rounded-bl-3xl lg:rounded-br-none ">
+          <p className="font-AveriaSerifLibre">
+            <span className="text-4xl">&quot;</span>
+            <span className="text-xl ">{character.quote}</span>
+            <span className="text-4xl leading-5">&quot;</span>
+          </p>
+        </div>
+      </motion.div>
+    </div>
   )
 }