+++ pages/videos/index.tsx
@@ -55,6 +55,14 @@ export default function VideosHome({ videos, theme }) {
           </div>
         </div>
         <motion.div className="w-full h-1/2">
+          <div className="relative z-50 bg-wall  py-2 mt-10 w-5/6 md:w-1/2 mx-auto shadow-m border-jasmine-faded border- ">
+            <p className="text-lg text-black text-center font-NotoSerif px-2">
+              The Meghan, Harry and Boris impressions came about via whimsical
+              enactment: I was with a friend, walking our dogs on a golf course
+              in Sunningdale, Berkshire, when I pretended to conduct a mock
+              interview: the premise was born.
+            </p>
+          </div>
           <div className="mt-20">
             <VideosRow
               videos={videos.royalInterview}